+++ chemprop/data/v2/__init__.py
@@ -1,4 +1,4 @@
 from .dataloader import MoleculeDataLoader
 from .datapoints import MoleculeDatapoint
-from .molecule import MoleculeDataset
+from .dataset import MoleculeDataset
 from .samplers import ClassBalanceSampler, SeededSampler
